You have an array of music genres:
const music = ["rock", "pop", "hip-hop", "electronic"];
Create a function myFavoriteMusic that uses the forEach method to display the text "My favorite music is {A} choosing from {B}" via the console.log for each element of the array. {A} is the current value of the element, {B} is the array itself.